requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.16.1
fsspec==2024.10.0
grpcio==1.67.0
grpcio-tools==1.67.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
hyperframe==6.0.1
idna==3.10
Jinja2==3.1.4
jiter==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.13
langchain-text-splitters==0.3.0
langsmith==0.1.137
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
openai==1.52.2
orjson==3.10.10
packaging==24.1
portalocker==2.10.1
protobuf==5.28.3
pydantic==2.9.2
pydantic_core==2.23.4
PyPDF2==3.0.1
PyYAML==6.0.2
qdrant-client==1.12.0
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.4.5
sniffio==1.3.1
sympy==1.13.1
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.20.1
torch==2.5.0
tqdm==4.66.6
transformers==4.45.2
typing_extensions==4.12.2
urllib3==2.2.3
google-generative-ai==0.1.0

